backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis & Celery
redis==5.2.1
celery[redis]==5.4.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# WebSocket
python-socketio==5.12.1

# AI
anthropic==0.43.0
httpx==0.28.1

# Monitoring
sentry-sdk[fastapi]==2.19.2

# Utilities
python-dateutil==2.9.0
